progress commit

+++ src/com/skilldistillery/jets/FighterJet.java
@@ -4,12 +4,12 @@ public class FighterJet extends Jet implements CombatReady {
 
 	public FighterJet() {
 		super();
-		
+
 	}
 
-	public FighterJet(String model, int speed, int range, int price) {
+	public FighterJet(String model, int speed, int range, double price) {
 		super(model, speed, range, price);
-		
+
 	}
 
 	public void dogFight() {
